requirements.txt
Django==4.2.6
djangorestframework==3.14.0
drf-nested-routers==0.93.4